ajustando relatorio aplicação

+++ backoffice/Application/DTOs/Cadastros/AplicacaoRecomendacoesTecnicas/ViewModel/AplicacaoRecomendacoesTecnicasViewModel.cs
@@ -13,7 +13,7 @@ public class AplicacaoRecomendacoesTecnicasViewModel
     public int? IdAeronave { get; set; }
     public int? IdAlturaVoo { get; set; }
     public string AlturaVooCustom { get; set; }
-    public int? Temperatura { get; set; }
+    public double? Temperatura { get; set; }
     public int? UrDoAR { get; set; }
     public int? VelocidadeVento { get; set; }
     public int? IdTipoDeProduto { get; set; }
